fix: importing server-only action to client
